Grpc: Use eventloop for server calls to avoid hanging
